HDPI-698: work in progress
