+++ src/main/steps/page3No/content.ts
@@ -1,7 +1,7 @@
 export const generateContent = () => {
   return {
-    title: 'You selected YES',
-    next: 'Continue',
+    title: 'You selected NO',
+    message: 'You cannot continue past this point.',
     backUrl: '/steps/page2'
   };
 };
