+++ src/main/steps/page3No/get.ts
@@ -1,7 +1,7 @@
 import { GetController } from '../../app/controllers/GetController';
 import { generateContent } from './content';
 
-export default class Page3YesGetController extends GetController {
+export default class Page3NoGetController extends GetController {
   constructor(view: string) {
     super(view, generateContent);
   }
