+++ src/main/steps/page3No/post.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 
-export default class Page3YesPostController {
+export default class Page3NoPostController {
   post = (req: Request, res: Response): void => {
     res.redirect('/steps/page4');
   };
