+++ src/main/steps/page3Yes/content.ts
@@ -1,7 +1,7 @@
 export const generateContent = () => {
   return {
-    title: 'You selected NO',
-    message: 'You cannot continue past this point.',
+    title: 'You selected YES',
+    next: 'Continue',
     backUrl: '/steps/page2'
   };
 };
