+++ src/main/steps/page3Yes/index.ts
@@ -1,14 +1,14 @@
-import Page3NoGetController from './get';
-import Page3NoPostController from './post';
 import { generateContent } from './content';
+import Page3YesGetController from './get';
+import Page3YesPostController from './post';
 
 export const step = {
-  url: '/steps/page3/no',
-  name: 'page3No',
-  view: 'steps/page3No/template.njk',
+  url: '/steps/page3/yes',
+  name: 'page3Yes',
+  view: 'steps/page3Yes/template.njk',
   generateContent,
   stepDir: __dirname,
-  getController: new Page3NoGetController('steps/page3No/template.njk'),
-  postController: new Page3NoPostController()
+  getController: new Page3YesGetController('steps/page3Yes/template.njk'),
+  postController: new Page3YesPostController()
 };
