+++ src/main/views/steps/page2/template.njk
@@ -3,8 +3,6 @@
 {% from "govuk/components/button/macro.njk" import govukButton %}
 {% from "govuk/components/back-link/macro.njk" import govukBackLink %}
 
-
-
 {% block content %}
   {{ govukBackLink({
     text: "Back",
