+++ src/main/views/template.njk
@@ -1,7 +1,6 @@
 {% extends "govuk/template.njk" %}
 {% from "govuk/components/header/macro.njk" import govukHeader %}
-
-
+{% from "govuk/components/phase-banner/macro.njk" import govukPhaseBanner %}
 
 {% block head %}
   <title>{% block pageTitle %}Possession claims{% endblock %} - GOV.UK</title>
@@ -9,17 +8,23 @@
 {% endblock %}
 
 {% block header %}
-{{ govukHeader({
-  navigation: [
-    {
-      text: 'logout',
-      href: '/logout'
-    }
-  ]
-}) }}
+  {{ govukHeader({
+    containerClasses: "govuk-width-container",
+    serviceName: serviceName or "Possession claims",
+    serviceUrl: "/"
+  }) }}
+  <div class="govuk-width-container">
+      {{ govukPhaseBanner({
+        tag: {
+          text: "Alpha"
+        },
+        html: 'This is a new service – your <a class="govuk-link" href="#">feedback</a> will help us to improve it.'
+      }) }}
+  </div>
 {% endblock %}
 
 {% block bodyEnd %}
+
   {# Run JavaScript at end of the <body>, to avoid blocking the initial render. #}
   {% include "webpack/js.njk" %}
 {% endblock %}
